rename sanity to accessibility
