+++ cypress/README.md
@@ -25,7 +25,7 @@ Use scenario tags:
 4. `@[bug-number]` for example `@FE-1234` - use this tag after `@ignore` to explain why test is ignored.
 5. `@ignore` and `@[bug-number]` should be removed after bug fix
 6. `@build` for pipeline - each build is verified by those tests
-7. `@sanity` tests verifies accessibility violations
+7. `@accessibility` tests verifies accessibility violations
 
 ## Files structure
 ```
@@ -34,7 +34,7 @@ Use scenario tags:
 │ ├── fixture
 │ ├── features
 │ │   ├── build
-│ │   ├── sanity
+│ │   ├── accessibility
 │ │   └── regression
 │ ├── locators
 │ │   └── [component-name]
